# Re write the result set