--- a/axelrod/__init__.py
+++ b/axelrod/__init__.py
@@ -18,7 +18,7 @@ from .strategies import *
 from .deterministic_cache import DeterministicCache
 from .match_generator import *
 from .tournament import Tournament
-from .result_set import ResultSet, ResultSetFromFile
+from .result_set import ResultSet
 from .ecosystem import Ecosystem
 from .fingerprint import AshlockFingerprint, TransitiveFingerprint