--- a/docs/conf.py
+++ b/docs/conf.py
@@ -21,7 +21,9 @@ MOCK_MODULES = [
     'scipy', 'scipy.stats','numpy', 'numpy.linalg', 'numpy.random',
     'matplotlib.pyplot', 'matplotlib','matplotlib.transforms', 'tqdm',
     'mpl_toolkits.axes_grid1', 'dill', 'multiprocess','prompt_toolkit',
-    'prompt_toolkit.token', 'prompt_toolkit.styles','prompt_toolkit.validation']
+    'prompt_toolkit.token', 'prompt_toolkit.styles','prompt_toolkit.validation',
+    'pandas', 'pandas.util', 'pandas.util.decorators', 'toolz', 'toolz.curried',
+    'toolz.functoolz', 'cloudpickle', 'dask', 'dask.dataframe']
 for mod_name in MOCK_MODULES:
     sys.modules[mod_name] = mock.Mock()