… axelrod/tests/unit/test_plot.py
… axelrod/tests/unit/test_resultset.py
… docs/tutorials/advanced/tournament_results.rst
… docs/tutorials/further_topics/morality_metrics.rst
… docs/tutorials/further_topics/spatial_tournaments.rst
… docs/tutorials/getting_started/index.rst
… docs/tutorials/getting_started/interactions.rst
… docs/tutorials/getting_started/summarising_tournaments.rst
… requirements.txt
… test_outputs/expected_test_tournament.csv
… test_outputs/expected_test_tournament_no_results.csv
… test_outputs/test_results.csv
… test_outputs/test_results_spatial.csv
… test_outputs/test_results_spatial_three.csv
… test_outputs/test_results_spatial_two.csv