**Note this PR is to a `possible-v4` branch which is a branch of `master`**

(@marcharper @meatballs Genuinely sorry for the diff on `result_set.py` it's a mess as this kind of guts that file... If we like the idea of what I've done here I'm happy to fix that as best I can.)

At present the result set plays the matches in parallel followed by a
(sometimes computationally expensive) single process of reading in and
analysing the interactions.

TLDR: This changes how the internal result set calculations are being
done. They are more efficiently calculated.

This happens by doing the following:

1. The various match by match calculations are done by the tournament
   (the winner of a match, the cooperation count, the score etc...). Mainly thanks to [`Tournament._calculate_results`](https://github.com/Axelrod-Python/Axelrod/pull/1165/files#diff-dad184e860a161b8b929a7ebbfd21c9cR412) 
2. These calculations are written to file (at present we just write the
  actual actions of an interaction) thanks to this line of [`Tournament._write_interaction_to_file`](https://github.com/Axelrod-Python/Axelrod/pull/1165/files#diff-dad184e860a161b8b929a7ebbfd21c9cR244)
3. The form of this file has also changed: for every match there are 2
   rows. One row corresponds to each player. This might seem more costly storage
   wise but is done to enable faster analysis (more on that later).
4. The analysis is now done entirely using `dask`: "Dask is a flexible
   parallel computing library for analytic computing."
   (https://dask.pydata.org/). This ensures all calculations are done on
   disk (so no huge memory problems) but also that they can be done **in
   parallel**. This is all done using the nice Pandas-like API that dask
   has so essentially all calculations for the result set are just done
   by a few `groupby` statements. The tasks to be run are built with [`ResultSet._build_tasks`](https://github.com/Axelrod-Python/Axelrod/pull/1165/files#diff-ba927cc36dbf4d89036c3602ad830638R535) and then compute (in parallel) with [`ResultSet._compute_tasks`](https://github.com/Axelrod-Python/Axelrod/pull/1165/files#diff-ba927cc36dbf4d89036c3602ad830638R525).
5. There is *some* work being done outside of dask but that's just
   reshaping data. `dask` outputs `pandas.Series` and to be consistent
   with our current setup these are changes to be lists of list etc... This is all done in [`ResultSet._reshape_out`](https://github.com/Axelrod-Python/Axelrod/pull/1165/files#diff-ba927cc36dbf4d89036c3602ad830638R80)

**Some user facing changes** (which is why I suggest this would be for a
`v4.0.0` release):

- The `result_set.interactions` is no longer possible. This is a choice
  and not forced by the redesign: I don't think this is ever necessary or
  helpful. The data file can always be read in.
- The ability to `tournament.play(in_memory=True)` has been removed.
  Again, not entirely a forced change (although it would be a tiny bit
  of work to allow this). Given the recent work to make everything work
  on Windows I don't think this is necessary and has allowed for a big
  deletion of code (which is a good thing re maintenance).
- The interactions data file is now in a different format, this is
  forced by the design choice.
- I have made a slight modification to `result_set.cooperation`.
  Currently this sets all self interactions to be 0 but I think that's not
  the right way to display it (note that the cooperation rates were all
  being done correctly).

**As well as ensuring the work done in series is reduced and the parallel
workers also calculate the scores (which I think is more efficient)**
this also seems to be faster:

On this branch:

```python
import axelrod as axl
players = [s() for s in axl.strategies if s.classifier["memory_depth"] == 1]
tournament = axl.Tournament(players, turns=200, repetitions=100)
results = tournament.play(processes=4)
```

Took: 1min 49s

```python
import axelrod as axl
players = [s() for s in axl.short_run_time_strategies]
tournament = axl.Tournament(players, turns=200, repetitions=20)
results = tournament.play(processes=4)
```

Took: 21min 2s

On current master;

```python
import axelrod as axl
players = [s() for s in axl.strategies if s.classifier["memory_depth"] == 1]
tournament = axl.Tournament(players, turns=200, repetitions=100)
results = tournament.play(processes=4)
```

Took: 2min 36s

```python
import axelrod as axl
players = [s() for s in axl.short_run_time_strategies]
tournament = axl.Tournament(players, turns=200, repetitions=20)
results = tournament.play(processes=4)
```

Took: 28min 8s

**One final aspect to consider** (I think) is that adding `dask` as a
dependency open up the potential to use it's `delayed` decorator to do
all our parallel processing. This would have the benefit of being able
to use a distributed scheduler that `dask` has. (We'd have to
investigate if this actually works based on our parallelisation but at
least the possibility is there).