@drvinceknight 2018-02-01
> All looks good to me. Nice piece of work!

Great, thanks for taking the time @meatballs :+1:

Once we're happy with this (@marcharper please don't hesitate to point out anything missing), we can merge and then open a PR from the `possible-v4` in to `master`. Not sure why I felt the need for an intermediate branch...